+++ Mods ChoGGi/Notes/metadata.lua
@@ -1,0 +1,32 @@
+return PlaceObj("ModDef", {
+	"dependencies", {
+		PlaceObj("ModDependency", {
+			"id", "ChoGGi_Library",
+			"title", "ChoGGi's Library",
+			"version_major", 10,
+			"version_minor", 6,
+		}),
+	},
+	"title", "Notes",
+	"id", "ChoGGi_Notes",
+	"steam_id", "2692788508",
+	"pops_any_uuid", "a754726f-1e92-450a-b2ae-61815e7d9cd0",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagTools", true,
+	"description", [[
+Places "note" buildings you can use to store text.
+Notepad opens when you select it, toggle in building menu.
+Mod options for width/height of notepad.
+
+Requested by samirahope.
+]],
+})